# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudyWave Q&A Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List all questions
      parameters:
        - name: subject
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Paginated questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedQuestions"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question with answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question with answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /questions/{id}/vote:
    post:
      operationId: voteQuestion
      tags: [questions]
      summary: Vote on a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Question:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        subject:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        authorDisplayName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        upvotes:
          type: integer
        downvotes:
          type: integer
        answerCount:
          type: integer
        hasAwardedAnswer:
          type: boolean
        isSolved:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - content
        - subject
        - authorId
        - authorUsername
        - authorDisplayName
        - upvotes
        - downvotes
        - answerCount
        - hasAwardedAnswer
        - isSolved
        - createdAt
        - updatedAt
    QuestionDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        subject:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        authorDisplayName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        authorPoints:
          type: integer
        upvotes:
          type: integer
        downvotes:
          type: integer
        hasAwardedAnswer:
          type: boolean
        isSolved:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
      required:
        - id
        - title
        - content
        - subject
        - authorId
        - authorUsername
        - authorDisplayName
        - authorPoints
        - upvotes
        - downvotes
        - hasAwardedAnswer
        - isSolved
        - createdAt
        - updatedAt
        - answers
    PaginatedQuestions:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - questions
        - total
        - page
        - limit
        - totalPages
    CreateQuestionBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        subject:
          type: string
      required:
        - title
        - content
        - subject
    UpdateQuestionBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        subject:
          type: string
    VoteBody:
      type: object
      properties:
        type:
          type: string
      required:
        - type
    Answer:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        questionId:
          type: integer
        authorId:
          type: integer
        authorUsername:
          type: string
        authorDisplayName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        authorPoints:
          type: integer
        upvotes:
          type: integer
        downvotes:
          type: integer
        isAwarded:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - content
        - questionId
        - authorId
        - authorUsername
        - authorDisplayName
        - authorPoints
        - upvotes
        - downvotes
        - isAwarded
        - createdAt
        - updatedAt
...