# lib/api-zod/src/generated/types/createQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyWave Q&A Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQuestionBody {
  title: string;
  content: string;
  subject: string;
}